requirements.txt
# Core image processing
Pillow>=10.0.0
piexif>=1.1.3
opencv-python>=4.8.0

# Perceptual hashing for advanced analysis
imagehash>=4.3.1

# ML/AI capabilities
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Metadata handling
iptcinfo3>=2.1.4
python-xmp-toolkit>=2.0.1; platform_system != "Windows"

# Utilities
click>=8.1.0
PyYAML>=6.0
tqdm>=4.65.0
python-dateutil>=2.8.2

# Image format support
pillow-heif>=0.13.0  # HEIC support
rawpy>=0.18.0  # RAW format support

# Video processing
moviepy>=1.0.3  # Video I/O and basic editing
scenedetect[opencv]>=0.6.1  # Scene detection and segmentation
ffmpeg-python>=0.2.0  # FFmpeg wrapper for advanced operations
pymediainfo>=6.0.1  # Video metadata extraction

# NVIDIA Build API integration
requests>=2.31.0  # API calls to NVIDIA Build
httpx>=0.25.0  # Async API client

# Advanced video analysis
# TensorRT optimization (install separately via NVIDIA)
# nvidia-pyindex>=1.0.9
# onnx>=1.15.0  # For TensorRT model conversion

# Optional: Face detection
# face-recognition>=1.3.0
# dlib>=19.24.0
